+++ e2e-playwright/page-object/components/material/index.ts
@@ -15,14 +15,5 @@
  * limitations under the License.
  */
 
-import { Page } from '@playwright/test';
-import { BaseComponent } from '../base.component';
-
-export class ErrorComponent extends BaseComponent {
-    private static rootElement = 'mat-error';
-    public content = this.getChild('');
-
-    constructor(page: Page) {
-        super(page, ErrorComponent.rootElement);
-    }
-}
+export * from './error.component';
+export * from './validation.component';
