+++ e2e-playwright/page-object/components/material/validation.component.ts
@@ -17,9 +17,10 @@
 
 import { Page } from '@playwright/test';
 import { BaseComponent } from '../base.component';
+import { materialLocators } from './material-locators';
 
 export class TooltipComponent extends BaseComponent {
-    private static rootElement = 'mat-tooltip-component';
+    private static rootElement = materialLocators.Tooltip.root;
     public content = this.getChild('div');
 
     constructor(page: Page) {
